Use the native git repo instead of a pkg manager
We used to rely on a package manager (`apt-get` or `brew`) to install
our `libatlasclient` dependency. The main problem with this approach
is the fact that those package managers will install the latest
available artifact, as opposed to the one used when we tested and built
this module. Now we just fetch a particular tagged release from github
and go from there.