Update to native client 3.7.0
Update to atlas native client v3.7.0 which uses the new config API.

For some reason one of the bucket timer test cases was failing, and it
was due to the test case being broken. It was getting a counter
with certain tags, and not a timer, which is what gets generated by
bucket timers. In a future PR we will have to address what happens when
the user asks for a meter with the same ID but a different type to one
already registered.